Rearrange some imports so that importing extra.utils doesn't import extra.components
